V0
Started with Directory structure and got the base of Flask working. Commands will be passed using subprocess and will be predefined. An API script will be used to manage getting the data to the Flask instance.

I feel it's worth pointing out that I am just messing around with this for fun and have very little skill!