V0.0.3
Added try except for reading from the config file, in case it errors because they aren't present.

Graph function added for todays 5 minutes chunks. It's very slow thought so could be sped up a bit with threading... that's a headake for later though. A graph shows on the HTML but has no CSS for it so it looks massive. Kinda done with the day with this now too lol...have more important things that need to be done